+++ src/rars/riscv/instructions/LB.java
@@ -36,7 +36,7 @@ public class LB extends Load {
     }
 
     public long load(int address) throws AddressErrorException {
-        return (Globals.memory.getByte(address) << 24) >> 24; // Shifting sign extends
+        return ((long) Globals.memory.getByte(address) << 24) >> 24; // Shifting sign extends
     }
 }
