+++ src/rars/riscv/instructions/LUI.java
@@ -1,9 +1,9 @@
 package rars.riscv.instructions;
 
 import rars.ProgramStatement;
-import rars.riscv.hardware.RegisterFile;
 import rars.riscv.BasicInstruction;
 import rars.riscv.BasicInstructionFormat;
+import rars.riscv.hardware.RegisterFile;
 
 /*
 Copyright (c) 2017,  Benjamin Landers
@@ -40,6 +40,6 @@ public class LUI extends BasicInstruction {
 
     public void simulate(ProgramStatement statement) {
         int[] operands = statement.getOperands();
-        RegisterFile.updateRegister(operands[0], operands[1] << 12);
+        RegisterFile.updateRegister(operands[0], (long) operands[1] << 12);
     }
 }
