+++ src/rars/riscv/instructions/SLLI.java
@@ -1,9 +1,9 @@
 package rars.riscv.instructions;
 
 import rars.ProgramStatement;
-import rars.riscv.hardware.RegisterFile;
 import rars.riscv.BasicInstruction;
 import rars.riscv.BasicInstructionFormat;
+import rars.riscv.hardware.RegisterFile;
 
 /*
 Copyright (c) 2017,  Benjamin Landers
@@ -36,11 +36,11 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 public class SLLI extends BasicInstruction {
     public SLLI() {
         super("slli t1,t2,10", "Shift left logical : Set t1 to result of shifting t2 left by number of bits specified by immediate",
-                BasicInstructionFormat.R_FORMAT, "0000000 ttttt sssss 001 fffff 0010011",false);
+                BasicInstructionFormat.R_FORMAT, "0000000 ttttt sssss 001 fffff 0010011", false);
     }
 
     public void simulate(ProgramStatement statement) {
         int[] operands = statement.getOperands();
-        RegisterFile.updateRegister(operands[0], RegisterFile.getValue(operands[1]) << operands[2]);
+        RegisterFile.updateRegister(operands[0], (long) RegisterFile.getValue(operands[1]) << operands[2]);
     }
 }
